Provide a common DatasetFactory base class
