+++ xarray_ms/backend/msv2/factories/antenna.py
@@ -1,32 +1,15 @@
-from typing import Dict, Mapping
-
 import numpy as np
 from xarray import Dataset, Variable
 
+from xarray_ms.backend.msv2.factories.core import DatasetFactory
 from xarray_ms.backend.msv2.imputation import maybe_impute_observation_table
-from xarray_ms.backend.msv2.structure import MSv2StructureFactory, PartitionKeyT
 from xarray_ms.errors import InvalidMeasurementSet
-from xarray_ms.multiton import Multiton
 
 RELOCATABLE_ARRAY = {"ALMA", "VLA", "NOEMA", "EVLA"}
 
 
-class AntennaDatasetFactory:
-  _partition_key: PartitionKeyT
-  _structure_factory: MSv2StructureFactory
-  _subtable_factories: Dict[str, Multiton]
-
-  def __init__(
-    self,
-    partition_key: PartitionKeyT,
-    structure_factory: MSv2StructureFactory,
-    subtable_factories: Dict[str, Multiton],
-  ):
-    self._partition_key = partition_key
-    self._structure_factory = structure_factory
-    self._subtable_factories = subtable_factories
-
-  def get_dataset(self) -> Mapping[str, Variable]:
+class AntennaDatasetFactory(DatasetFactory):
+  def get_dataset(self) -> Dataset:
     partition = self._structure_factory.instance[self._partition_key]
     ants = self._subtable_factories["ANTENNA"].instance
     feeds = self._subtable_factories["FEED"].instance
